LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-scroll-007-expected.txt


PASS #target3 is scroll-adjusted in both axises
